delegate notification dispatching from DownloadBatchStatusDispatcher to the generic.
